requirements.txt
fastapi
uvicorn
edge-tts
aiohttp
python-dotenv
